Improve memory allocation of lods with CityJSONWorkerParser

+++ src/parsers/CityJSONWorkerParser.js
@@ -168,7 +168,7 @@ export class CityJSONWorkerParser {
 			geom.setAttribute( 'geometryid', new BufferAttribute( geomIdsArray, 1 ) );
 			const lodIdsArray = new Float32Array( e.data.lodIds );
 			geom.setAttribute( 'lodid', new BufferAttribute( lodIdsArray, 1 ) );
-			const boundaryIdsArray = new Float32Array( e.data.boundaryIds );
+			const boundaryIdsArray = new Int8Array( e.data.boundaryIds );
 			geom.setAttribute( 'boundaryid', new BufferAttribute( boundaryIdsArray, 1 ) );
 
 			geom.attributes.position.needsUpdate = true;
